% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logic.R
\name{get_current_number}
\alias{get_current_number}
\title{Get the number at a given position in the sequence}
\usage{
get_current_number(sequence, index)
}
\arguments{
\item{sequence}{A numeric vector}

\item{index}{Position to retrieve (1-based)}
}
\value{
A single numeric value
}
\description{
Get the number at a given position in the sequence
}
\details{
Retrieves the number the player is currently shown. The index
  advances each turn, revealing one number at a time.
}
\examples{
seq <- c(42, 7, 198)
get_current_number(seq, 1)
get_current_number(seq, 2)
}
