% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logic.R
\name{generate_sequence}
\alias{generate_sequence}
\title{Generate a sequence of random numbers}
\usage{
generate_sequence(n = 3)
}
\arguments{
\item{n}{Number of values to draw}
}
\value{
A numeric vector of length n
}
\description{
Generate a sequence of random numbers
}
\details{
Samples n unique integers from 0 to 1000 without replacement.
}
\examples{
generate_sequence()
generate_sequence(n = 5)
}
